test(core): reorganize test wrappers
