+++ src/patch-place-break-core/src/test/java/fr/djaytan/mc/jrppb/core/config/ConfigApiTest.java
@@ -28,8 +28,8 @@ import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
 import fr.djaytan.mc.jrppb.commons.test.TestResourcesHelper;
 import fr.djaytan.mc.jrppb.core.config.serialization.ConfigSerializer;
-import fr.djaytan.mc.jrppb.core.config.testutils.ValidatorTestWrapper;
 import fr.djaytan.mc.jrppb.core.config.validation.PropertiesValidator;
+import fr.djaytan.mc.jrppb.core.config.validation.ValidatorTestWrapper;
 import java.io.IOException;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
