+++ src/patch-place-break-core/src/test/java/fr/djaytan/mc/jrppb/core/config/ConfigManagerTest.java
@@ -37,8 +37,8 @@ import fr.djaytan.mc.jrppb.core.config.properties.DbmsHostPropertiesImpl;
 import fr.djaytan.mc.jrppb.core.config.properties.DbmsServerPropertiesImpl;
 import fr.djaytan.mc.jrppb.core.config.properties.Properties;
 import fr.djaytan.mc.jrppb.core.config.serialization.ConfigSerializer;
-import fr.djaytan.mc.jrppb.core.config.testutils.ValidatorTestWrapper;
 import fr.djaytan.mc.jrppb.core.config.validation.PropertiesValidator;
+import fr.djaytan.mc.jrppb.core.config.validation.ValidatorTestWrapper;
 import fr.djaytan.mc.jrppb.core.storage.api.properties.DataSourceProperties;
 import fr.djaytan.mc.jrppb.core.storage.api.properties.DataSourceType;
 import java.io.IOException;
