+++ src/patch-place-break-core/src/test/java/fr/djaytan/mc/jrppb/core/config/properties/ConnectionPoolPropertiesImplTest.java
@@ -33,8 +33,8 @@ import com.google.common.jimfs.Jimfs;
 import com.jparams.verifier.tostring.NameStyle;
 import com.jparams.verifier.tostring.ToStringVerifier;
 import fr.djaytan.mc.jrppb.commons.test.TestResourcesHelper;
-import fr.djaytan.mc.jrppb.core.config.testutils.ConfigSerializerTestWrapper;
-import fr.djaytan.mc.jrppb.core.config.testutils.ValidatorTestWrapper;
+import fr.djaytan.mc.jrppb.core.config.serialization.ConfigSerializerTestWrapper;
+import fr.djaytan.mc.jrppb.core.config.validation.ValidatorTestWrapper;
 import jakarta.validation.ConstraintViolation;
 import java.io.IOException;
 import java.io.UncheckedIOException;
